% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_SMR.R
\name{mcmc_SMR}
\alias{mcmc_SMR}
\title{Fit the generalized categorical spatial mark resight model. SMR_move fits the activity centers to cameras and snares separately whereas SMR fits them concurrently.}
\usage{
mcmc_SMR(
  data,
  input = list(niter = 2400, nburn = 1200, nthin = 5, M = 200, mobile_center = F, inits =
    list(lam0_mark = 0.05, lam0_sight = 0.009, sigma_d = 0.1, sigma_p = 0.1, s1 = 5, s2 =
    5, psi = 0.5, gamma = vector("list", 1)), obstype = c("bernoulli", "poisson"), nswap
    = NA, proppars = list(lam0_mark = 0.05, lam0_sight = 0.1, sigma_d = 0.02, sigma_p =
    0.2, s1 = 0.5, s2 = 0.25, s2t = 0.1), max_proppars <- list(lam0_mark = 100,
    lam0_sight = 100, sigma_d = 100, sigma_p = 100, s1 = 100, s2 = 100, s2t = 100),
    min_proppars <- list(lam0_mark = 0.001, 
     lam0_sight = 0.001, sigma_d = 0.001,
    sigma_p = 0.001, s1 = 0.001, s2 = 0.001, s2t = 0.001), storeLatent = TRUE, storeGamma
    = TRUE, IDup = "Gibbs", tf1 = NA, tf2 = NA)
)
}
\arguments{
\item{data}{a data list}

\item{input}{Model definitions and inputs}
}
\description{
This function fits the generalized categorical spatial mark resight model.
Modelling the marking process relaxes the assumption that the distribution of marked
individuals across the landscape is spatially uniform.

the data list should all elements of that object are necessary.
y_mark, y_sight_marked, y_sight_unmarked, G_marked, and G_unmarked are necessary
list elements. y_sight_x and G_x for x=unk and marke.noID are necessary if there are samples
of unknown marked status or samples from marked samples without individual identities.

"niter" the number of MCMC iterations to perform
"nburn" the number of MCMC iterations to discard as burnin
"nthin" the MCMC thinning interval. Keep every nthin iterations.
"M" the level of data augmentation
"inits" a list of initial values for lam0_mark,lam0_sight, sigma, gamma, and psi.
The list element for gamma is itself a list with ncat elements. See the example below.
"obstype" a vector of length two indicating the observation model, "bernoulli" or "poisson", for the
marking and sighting process
"nswap" an integer indicating how many samples for which the latent identities
are updated on each iteration.
"proppars" a list of proposal distribution tuning parameters for lam0_mark,
lam0_sight, sigma, s, and st, for the activity centers of untelemetered and
telemetered individuals, respectively. The tuning parameter should be
smaller for individuals with telemetry and increasingly so as the number of
locations per individual increases
"storeLatent" a logical indicator for whether or not the posteriors of the
latent individual identities, z, and s are stored and returned
"storeGamma" a logical indicator for whether or not the posteriors for gamma are stored and returned
"IDup" a character string indicating whether the latent identity update is
done by Gibbs or Metropolis-Hastings, "Gibbs", or "MH". For
obstype="bernoulli", only "MH" is available because the full conditional is not known.
"tf1" a trap operation vector or matrix for the marking process. If exposure to capture does
not vary by individual, tf1 should be a vector of length J1 indicating how many of the K1 occasions
each marking location was operational. If exposure to capture varies by individual or by trap and
individual, tf1 should be a matrix of dimension M x J1 indicating how many
of the K1 occasions individual i was exposed to at trap j. This allows known
additions or removals during the marking process to be accounted for.
Exposure for the n_marked+1 ... M uncaptured individuals should be the
same as the number of occasions each trap was operational. We can't account for unknown
additions and removals.
"tf2" a trap operation vector or matrix for the sighting process. If exposure to capture does
not vary by indiviudal, tf2 should be a vector of length J2 indicating how many of the K2 occasions
each sighting location was operational. If exposure to capture varies by individual or by trap and
individual, tf2 should be a matrix of dimension M x J2 indicating how many
of the K2 occasions individual i was exposed to at trap j. This allows
known additions or removals between the marking and sighting processes and
during the sighting process to be accounted for. Exposure for
the n_marked+1 ... M uncaptured individuals should be the
same as the number of occasions each trap was operational. We can't account for unknown
additions and removals.
"X1" is a matrix of marking coordinates
"X2" is a matrix of sighting coordinates, ,
"K1", the integer number of marking occasions, and an element "K2",
the integer number of sighting occasions are necessary.
"IDlist" is a list containing elements ncat and IDcovs. ncat is an integer for the number
of categorical identity covariates and IDcovs is a list of length ncat with elements containing the
values each categorical identity covariate may take.
"locs" is an n_marked x nloc x  2 array of telemetry locations is optional. This array can
have missing values if not all individuals have the same number of locations
and the entry for individuals with no telemetry should all be missing values (coded NA).
"markedS" is required if marking and sighting sessions are interspersed. This is a
n_marked x K2 matrix with 0 indicating an individual was not marked on occasion k and 1 if it
was.
"lam0_mark" is the coefficient determining the probability of marking an individual
"lam0_sight" is the coefficient determining the probability of sighting an individual
"sigma_d" is the variance for sighting probability distributions
"sigma_p" is the variance for marking probability distributions
"gamma" is a list of the category level probabilities of the same dimension
as IDcovs. The category level probabilities for each covariate must sum to 1
"psi" is probability of keeping augmented activity centers
"s1" and "s2" are the activity centers for marked and sighted augmented data
"proppars$s" is variance for s
"proppars$st" is variance for s w/ telemetry data
}
\details{
Allowing for activity center relocation (mobile_center = "move") means the model
incorporates a process for updating activity centers between the marking and
sighting phases. If the data are sparse, there is no telemetry data, or if
the study period is short, it may better to not allow for activity center
relocation (mobile_center = "no move").
}
