% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdgt.secr.host.R
\docType{package}
\name{spdgt.secr.host-package}
\alias{spdgt.secr.host}
\alias{spdgt.secr.host-package}
\title{spdgt.secr.host: Capture-recapture Models for Host Detection}
\description{
Run mark-recapture models
}
\author{
\strong{Maintainer}: Guen Grosklos \email{guen.grosklos@speedgoat.io}

Authors:
\itemize{
  \item Josh Nowak \email{josh.nowak@speedgoat.io}
  \item Hans Martin \email{hans.martin@speedgoat.io}
}

}
